todo: check off phase 6 (chunk streaming v2); codec test lint

+++ crates/soils-protocol/src/chunk_codec.rs
@@ -213,7 +213,7 @@ mod tests {
                         16 => 1,
                         _ => 0,
                     };
-                    v.set(x, y as i32, z, id);
+                    v.set(x, y, z, id);
                 }
             }
         }
